backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
boto3==1.35.76
botocore==1.35.99
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.1.8
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
easyocr==1.7.2
exceptiongroup==1.3.0
fastapi==0.121.2
filelock==3.20.0
fonttools==4.60.1
fsspec==2025.10.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
ImageIO==2.37.2
Jinja2==3.1.6
jmespath==1.0.1
kiwisolver==1.4.9
lazy_loader==0.4
MarkupSafe==3.0.3
matplotlib==3.10.7
minio==7.2.18
mpmath==1.3.0
networkx==3.5
ninja==1.13.0
nodeenv==1.9.1
numpy==2.2.6
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
packaging==25.0
pillow==11.0.0
polars==1.35.2
polars-runtime-32==1.35.2
prisma==0.15.0
psutil==7.1.3
pyclipper==1.3.0.post6
pycparser==2.23
pycryptodome==3.23.0
pydantic==2.12.4
pydantic-settings==2.11.0
pydantic_core==2.41.5
PyMuPDF==1.26.6
pyparsing==3.2.5
pypdf==6.2.0
python-bidi==0.6.7
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
PyYAML==6.0.3
requests==2.32.5
s3transfer==0.10.4
scikit-image==0.25.2
scipy==1.16.3
setuptools==80.9.0
shapely==2.1.2
six==1.17.0
sniffio==1.3.1
starlette==0.49.3
StrEnum==0.4.15
sympy==1.14.0
tifffile==2025.10.16
tomlkit==0.13.3
torch==2.9.1
torchvision==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
ultralytics==8.3.228
ultralytics-thop==2.0.18
urllib3==2.5.0
uvicorn==0.38.0
minio>=7.2.0
PyMuPDF>=1.23.0
opencv-python>=4.8.0
pypdf>=5.0.0
ultralytics>=8.0.0
Pillow>=10.0.0
python-multipart>=0.0.9
